+++ src/ralph_assets/forms.py
@@ -66,7 +66,7 @@ asset_fieldset = lambda: OrderedDict([
     ('Basic Info', [
         'type', 'category', 'model', 'niw', 'barcode', 'sn', 'warehouse',
         'location', 'status', 'task_url', 'loan_end_date', 'remarks',
-        'service_name', 'property_of',
+        'service_name', 'property_of', 'region',
     ]),
     ('Financial Info', [
         'order_no', 'invoice_date', 'invoice_no', 'price', 'provider',
@@ -835,6 +835,7 @@ class BaseAddAssetForm(DependencyAssetForm, BaseAssetForm):
             'property_of',
             'provider',
             'provider_order_date',
+            'region',
             'remarks',
             'request_date',
             'required_support',
@@ -999,6 +1000,7 @@ class BaseEditAssetForm(DependencyAssetForm, BaseAssetForm):
             'property_of',
             'provider',
             'provider_order_date',
+            'region',
             'remarks',
             'request_date',
             'required_support',
@@ -1187,6 +1189,12 @@ class AddDeviceForm(BaseAddAssetForm, MultivalFieldForm):
         validators=[validate_imeis],
     )
 
+    def __init__(self, request, *args, **kwargs):
+        super(AddDeviceForm, self).__init__(*args, **kwargs)
+        self.request = request
+        regions = request.user.get_profile().get_regions()
+        self.fields['region'] = ModelChoiceField(queryset=regions)
+
     def clean(self):
         """
         These form requirements:
@@ -1284,7 +1292,7 @@ class EditPartForm(BaseEditAssetForm):
 
 class EditDeviceForm(BaseEditAssetForm):
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, request, *args, **kwargs):
         super(EditDeviceForm, self).__init__(*args, **kwargs)
         self.fieldsets = asset_fieldset()
         self.fieldsets['Assigned licenses info'] = ['licences']
@@ -1292,6 +1300,9 @@ class EditDeviceForm(BaseEditAssetForm):
             'required_support',
             'supports',
         ]
+        self.request = request
+        regions = request.user.get_profile().get_regions()
+        self.fields['region'] = ModelChoiceField(queryset=regions)
 
     def clean(self):
         cleaned_data = super(EditDeviceForm, self).clean()