+++ src/ralph_assets/tests/functional/tests_history.py
@@ -7,6 +7,7 @@ from __future__ import unicode_literals
 
 from django.test import TestCase
 from django.core.urlresolvers import reverse
+from ralph.account.models import Region
 
 from ralph.cmdb.tests.utils import CIRelationFactory
 from ralph_assets.models_assets import (
@@ -55,6 +56,7 @@ class HistoryAssetsView(TestCase):
             'price': 10,
             'property_of': self.owner.id,
             'provider': 'test_provider',
+            'region': Region.get_default_region().id,
             'remarks': 'test_remarks',
             'size': 1,
             'sn': '666-666-666',
@@ -160,6 +162,7 @@ class HistoryAssetsView(TestCase):
         form = response.context['asset_form']
         update_dict = form.__dict__['initial']
         update_dict.update({
+            'region': Region.get_default_region().id,
             'required_support': 1,
             'asset': True,
         })
@@ -192,6 +195,7 @@ class ConnectAssetWithDevice(TestCase):
             'price': 10,
             'production_year': 2011,
             'provider': 'test_provider',
+            'region': Region.get_default_region().id,
             'remarks': 'test_remarks',
             'service': ci_relation.parent.id,
             'source': 1,
@@ -271,6 +275,7 @@ class TestsStockDevice(TestCase):
             'order_no': 2,
             'price': 10,
             'provider': 'test_provider',
+            'region': Region.get_default_region().id,
             'remarks': 'test_remarks',
             'service': ci_relation.parent.id,
             'sn': 'fake-sn',